+++ semcore/popper/src/Popper.jsx
@@ -522,6 +522,7 @@ function PopperPopper(props) {
     animationsDisabled,
     popper,
     focusableTriggerReturnFocusToRef,
+    focusMaster = false,
   } = props;
   const ref = React.useRef(null);
 
@@ -533,6 +534,7 @@ function PopperPopper(props) {
     autoFocus,
     interaction === 'focus' ? focusableTriggerReturnFocusToRef : triggerRef,
     !visible || disableEnforceFocus,
+    focusMaster,
   );
 
   useContextTheme(ref, visible);