+++ semcore/dropdown-menu/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.14.0] - 2023-11-28
+
+### Added
+
+- `interaction='click'` by default. Pass `focusMaster=true` for `Popover` if `interaction='click'`.
+
+### Removed
+
+- unnecessary style for focused `Popper`.
+
 ## [4.13.0] - 2023-11-23
 
 ### Added