+++ semcore/dropdown/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.12.0] - 2023-11-28
+
+### Added
+
+- `interaction='click'` by default. Pass `focusMaster=true` for `Popover` if `interaction='click'`.
+
 ## [4.11.0] - 2023-11-24
 
 ### Changed